Extract UrgencyField
