+++ apps/back-office/src/app/melden/_components/index.ts
@@ -2,3 +2,4 @@ export { InvalidFormAlert } from './InvalidFormAlert'
 export { PrimaryField } from './PrimaryField/PrimaryField'
 export { SourceField } from './SourceField/SourceField'
 export { SystemErrorAlert } from './SystemErrorAlert/SystemErrorAlert'
+export { UrgencyField } from './UrgencyField/UrgencyField'
